Moved deployments to the deployment folder for better organization
